requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.12.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
frozenlist==1.5.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
idna==3.10
multidict==6.1.0
packaging==24.2
pillow==11.1.0
playwright==1.53.0
propcache==0.2.1
pyee==13.0.0
pytesseract==0.3.13
python-dotenv==1.2.1
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.3
yarl==1.18.3
